requirements.txt
Flask
Flask-Cors
gunicorn
requests
PyMuPDF
PyPDF2
Pillow
pytesseract
cloudinary